update swift5
* addXcodeProj
* add symLink to main.swift
* remove mold.swift
